[StyleCop]Fix all the warnings in the AI.Luis.Tests project - Reorder methods and fields - Create new file for a class - Add trailing comma - Remove spaces
