+++ tests/Microsoft.Bot.Builder.AI.LUIS.Tests/MockedHttpClientHandler.cs
@@ -1,0 +1,43 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Net.Http;
+using System.Reflection;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Bot.Builder.Adapters;
+using Microsoft.Bot.Configuration;
+using Microsoft.Bot.Schema;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+using RichardSzalay.MockHttp;
+
+namespace Microsoft.Bot.Builder.AI.Luis.Tests
+{
+    public class MockedHttpClientHandler : HttpClientHandler
+    {
+        private readonly HttpClient client;
+
+        public MockedHttpClientHandler(HttpClient client)
+        {
+            this.client = client;
+        }
+
+        protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+        {
+            var mockedRequest = new HttpRequestMessage()
+            {
+                RequestUri = request.RequestUri,
+                Content = request.Content,
+                Method = request.Method,
+            };
+            return client.SendAsync(mockedRequest, cancellationToken);
+        }
+    }
+}
